veglegesit

+++ app/Http/Controllers/HibakController.php
@@ -39,7 +39,7 @@ class HibakController
                 if (request()->has('auto_azonosito')) {
                     $query->where('auto_azonosito', 'like', "%" . request('auto_azonosito') . "%");
                 }
-          //      $query->whereRaw(" user_id IN (Select id from users where root_user=?) or user_id = ?", [\Auth::id(), \Auth::id()]);
+                $query->whereRaw(" user_id IN (Select root_user from users where id=?) or user_id = ?", [\Auth::id(), \Auth::id()]);
 
             }, true)
             ->toJson();
@@ -47,13 +47,13 @@ class HibakController
 
     public function create()
     {
-        $select = Autok::whereRaw(" rejtett = 0 and ( user_id IN (Select id from users where root_user=?) or user_id = ? )", [\Auth::id(), \Auth::id()])->get()->toArray();
+        $select = Autok::whereRaw(" rejtett = 0 and ( user_id IN (Select root_user from users where id=?) or user_id = ? )", [\Auth::id(), \Auth::id()])->get()->toArray();
         return view('hibak.formview', compact('select'));
     }
 
     public function update($azonosito)
     {
-        $select = Autok::whereRaw(" rejtett = 0 and ( user_id IN (Select id from users where root_user=?) or user_id = ? )", [\Auth::id(), \Auth::id()])->get()->toArray();
+        $select = Autok::whereRaw(" rejtett = 0 and ( user_id IN (Select root_user from users where id=?) or user_id = ? )", [\Auth::id(), \Auth::id()])->get()->toArray();
         $model = (array)DB::table('hibajegy')->where("azonosito", $azonosito)->get()->first();
         return view('hibak.formview', compact("model", "azonosito", "select"));
     }
@@ -127,7 +127,9 @@ class HibakController
 
     public function delete(Request $request)
     {
-        return DB::table('hibajegy')->where('azonosito', $request->post('azonosito'))->delete();
+        if(session('role', 0) == 3) {
+            return DB::table('hibajegy')->where('azonosito', $request->post('azonosito'))->delete();
+        }
     }
 
 }
